updated for linklist and dashboarddisplay
